+++ .gitignore
@@ -13,6 +13,7 @@ yarn-error.log*
 /.changelog
 .npm/
 yarn.lock
+firebase.js
