+++ src/App.js
@@ -1,6 +1,6 @@
 import React from "react";
 import Room from "./components/Room";
-import Chat from "./components/Chat";
+import Chat from "./components/Chat/Chat";
 
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
