+++ src/components/Chat/Chat.js
@@ -1,6 +1,6 @@
 import React from "react";
-import Header from "./Header";
-import Navigation from "./Navigation";
+import Header from "../common/Header";
+import Navigation from "../common/Navigation";
 
 const Chat = () => {
   return (
